refactor: deduplicate shared code across controllers (#118)
Extract shared helpers (isSecretReady, createOrUpdateSecret) into
helpers.go, reuse caHTTPClient() in fetchCRL(), and split two large
controller files into logical sub-files for maintainability.

No logic changes — pure code movement and helper extraction.
